requirements.txt
# Requirements to deploy this code. To install all these run the following from terminal: 
# pip install -r requirements.txt

Django==1.5.4
Mezzanine==1.4.15
Pillow==2.1.0
bleach==1.2.2
django-browserid==0.9
django-simple-history==1.3.0
fancy-tag==0.2.0
filebrowser-safe==0.2.28
grappelli-safe==0.2.22
html5lib==0.95
oauthlib==0.6.0
pytz==2013d
requests==1.2.3
requests-oauthlib==0.3.3
six==1.4.1
wsgiref==0.1.2
